# openapi.json
{
  "components": {
    "schemas": {
      "Author": {
        "additionalProperties": false,
        "description": "Basic information about a Moxfield user.",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "profile_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Image Url"
          },
          "user_name": {
            "title": "User Name",
            "type": "string"
          }
        },
        "required": [
          "user_name"
        ],
        "title": "Author",
        "type": "object"
      },
      "DeckBoard": {
        "additionalProperties": false,
        "description": "A board (mainboard, sideboard, commanders, etc.) within a deck.",
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/DeckCard"
            },
            "title": "Cards",
            "type": "array"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "DeckBoard",
        "type": "object"
      },
      "DeckCard": {
        "additionalProperties": false,
        "description": "A single card entry within a board.",
        "properties": {
          "card": {
            "additionalProperties": true,
            "title": "Card",
            "type": "object"
          },
          "finish": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish"
          },
          "is_alter": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Alter"
          },
          "is_foil": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Foil"
          },
          "is_proxy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Proxy"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "quantity"
        ],
        "title": "DeckCard",
        "type": "object"
      },
      "DeckDetail-Input": {
        "additionalProperties": false,
        "description": "Full deck details including card breakdown.",
        "properties": {
          "authors": {
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "title": "Authors",
            "type": "array"
          },
          "boards": {
            "items": {
              "$ref": "#/components/schemas/DeckBoard"
            },
            "title": "Boards",
            "type": "array"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "title": "Color Identity",
            "type": "array"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "title": "Colors",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Author"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Description"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "hubs": {
            "items": {
              "type": "string"
            },
            "title": "Hubs",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "last_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "public_url": {
            "title": "Public Url",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/DeckStats"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/DeckTag"
            },
            "title": "Tags",
            "type": "array"
          },
          "tokens": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tokens",
            "type": "array"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          }
        },
        "required": [
          "public_id",
          "name",
          "format",
          "public_url"
        ],
        "title": "DeckDetail",
        "type": "object"
      },
      "DeckDetail-Output": {
        "additionalProperties": false,
        "description": "Full deck details including card breakdown.",
        "properties": {
          "authors": {
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "title": "Authors",
            "type": "array"
          },
          "boards": {
            "items": {
              "$ref": "#/components/schemas/DeckBoard"
            },
            "title": "Boards",
            "type": "array"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "title": "Color Identity",
            "type": "array"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "title": "Colors",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Author"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Description"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "hubs": {
            "items": {
              "type": "string"
            },
            "title": "Hubs",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "last_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "public_url": {
            "title": "Public Url",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/DeckStats"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/DeckTag"
            },
            "title": "Tags",
            "type": "array"
          },
          "tokens": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tokens",
            "type": "array"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          }
        },
        "required": [
          "public_id",
          "name",
          "format",
          "public_url"
        ],
        "title": "DeckDetail",
        "type": "object"
      },
      "DeckStats": {
        "additionalProperties": false,
        "description": "Aggregate deck statistics.",
        "properties": {
          "bookmark_count": {
            "default": 0,
            "title": "Bookmark Count",
            "type": "integer"
          },
          "comment_count": {
            "default": 0,
            "title": "Comment Count",
            "type": "integer"
          },
          "like_count": {
            "default": 0,
            "title": "Like Count",
            "type": "integer"
          },
          "view_count": {
            "default": 0,
            "title": "View Count",
            "type": "integer"
          }
        },
        "title": "DeckStats",
        "type": "object"
      },
      "DeckSummary": {
        "additionalProperties": false,
        "description": "Lightweight deck metadata that excludes card lists.",
        "properties": {
          "authors": {
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "title": "Authors",
            "type": "array"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "title": "Color Identity",
            "type": "array"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "title": "Colors",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Author"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Description"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "hubs": {
            "items": {
              "type": "string"
            },
            "title": "Hubs",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "last_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "public_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Url"
          },
          "stats": {
            "$ref": "#/components/schemas/DeckStats"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/DeckTag"
            },
            "title": "Tags",
            "type": "array"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          }
        },
        "required": [
          "public_id",
          "name",
          "format"
        ],
        "title": "DeckSummary",
        "type": "object"
      },
      "DeckTag": {
        "additionalProperties": false,
        "description": "Tags assigned by the deck author to individual cards.",
        "properties": {
          "card_name": {
            "title": "Card Name",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "card_name"
        ],
        "title": "DeckTag",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "UserDeckSummariesResponse": {
        "additionalProperties": false,
        "description": "Response for routes returning decks without card details.",
        "properties": {
          "decks": {
            "items": {
              "$ref": "#/components/schemas/DeckSummary"
            },
            "title": "Decks",
            "type": "array"
          },
          "total_decks": {
            "title": "Total Decks",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "required": [
          "user",
          "total_decks"
        ],
        "title": "UserDeckSummariesResponse",
        "type": "object"
      },
      "UserDecksResponse": {
        "additionalProperties": false,
        "description": "Main response payload for the /users/{username}/decks route.",
        "properties": {
          "decks": {
            "items": {
              "$ref": "#/components/schemas/DeckDetail-Output"
            },
            "title": "Decks",
            "type": "array"
          },
          "total_decks": {
            "title": "Total Decks",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "required": [
          "user",
          "total_decks"
        ],
        "title": "UserDecksResponse",
        "type": "object"
      },
      "UserSummary": {
        "additionalProperties": false,
        "description": "Top-level user descriptor returned by the API.",
        "properties": {
          "badges": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Badges",
            "type": "array"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "profile_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Image Url"
          },
          "profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          },
          "user_name": {
            "title": "User Name",
            "type": "string"
          }
        },
        "required": [
          "user_name"
        ],
        "title": "UserSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Simple proxy API that fetches public Moxfield data.",
    "title": "Moxfield Scraping API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Useful for uptime checks.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Check Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Check",
        "tags": [
          "meta"
        ]
      }
    },
    "/users/{username}/deck-summaries": {
      "get": {
        "operationId": "get_user_deck_summaries_users__username__deck_summaries_get",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeckSummariesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch a user's decks without card details.",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{username}/decks": {
      "get": {
        "operationId": "get_user_decks_users__username__decks_get",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDecksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch a user's decks including full card details.",
        "tags": [
          "users"
        ]
      }
    }
  }
}